#4 Feat mobile 📱  responsive 🧐 (#135)

+++ components/SectionsColumn.js
@@ -2,17 +2,13 @@ import {
   closestCenter,
   DndContext,
   KeyboardSensor,
-  PointerSensor,
+  MouseSensor,
+  TouchSensor,
   useSensor,
   useSensors,
 } from '@dnd-kit/core'
 import { restrictToVerticalAxis } from '@dnd-kit/modifiers'
-import {
-  arrayMove,
-  SortableContext,
-  sortableKeyboardCoordinates,
-  verticalListSortingStrategy,
-} from '@dnd-kit/sortable'
+import { arrayMove, SortableContext, sortableKeyboardCoordinates } from '@dnd-kit/sortable'
 
 import { SortableItem } from './SortableItem'
 import { useTranslation } from 'next-i18next'
@@ -28,7 +24,8 @@ export const SectionsColumn = ({
   getTemplate,
 }) => {
   const sensors = useSensors(
-    useSensor(PointerSensor),
+    useSensor(MouseSensor),
+    useSensor(TouchSensor),
     useSensor(KeyboardSensor, {
       coordinateGetter: sortableKeyboardCoordinates,
     })
@@ -97,17 +94,21 @@ export const SectionsColumn = ({
   }
 
   const resetSelectedSections = () => {
-    const data = localStorage.getItem('current-slug-list')
-
-    const sectionResetConfirmed = window.confirm(
-      'All sections of your readme will be removed; to continue, click OK'
-    )
-    if (sectionResetConfirmed === true) {
-      const slugList = data ? data.split(',') : []
-      setSectionSlugs((prev) => [...prev, ...slugList].filter((s) => s !== 'title-and-description'))
-      setSelectedSectionSlugs(['title-and-description'])
-      setFocusedSectionSlug('title-and-description')
-      localStorage.setItem('current-focused-slug', 'noEdit')
+    let data = localStorage.getItem('current-slug-list')
+    if (data) {
+      const sectionResetConfirmed = window.confirm(
+        'All sections of your readme will be removed; to continue, click OK'
+      )
+      if (sectionResetConfirmed === true) {
+        let slugList = []
+        slugList = localStorage.getItem('current-slug-list').split(',')
+        slugList.forEach((entry) => {
+          setSectionSlugs((prev) => prev.filter((s) => s !== 'title-and-description'))
+        })
+        setSelectedSectionSlugs(['title-and-description'])
+        setFocusedSectionSlug('title-and-description')
+        localStorage.setItem('current-focused-slug', 'noEdit')
+      }
     }
   }
